requirements.txt
# Core
streamlit>=1.40.0
python-dotenv>=1.0.1

# LangChain ecosystem (modern LCEL-based)
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0

# Vector store
chromadb>=0.5.0

# OpenAI SDK
openai>=1.50.0

# Document loaders
pypdf>=5.0.0
python-docx>=1.1.0
docx2txt>=0.8
wikipedia>=1.4.0

# Hybrid search (BM25 keyword retrieval)
rank_bm25>=0.2.2

# Numerical (semantic chunking cosine similarity)
numpy>=1.24.0

